restructured files and removed unnecessary prints from AddToFavourites.py
